pyproject.toml
[project]
name = "ai-support-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "google-cloud-aiplatform>=1.147.0",
    "google-genai>=1.72.0",
    "loguru>=0.7.3",
    "numpy>=2.4.4",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
